+++ src/main/java/kr/hs/entrydsm/rollsroyce/domain/admin/presentation/AdminController.java
@@ -7,22 +7,31 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
+import javax.validation.Valid;
+
+import kr.hs.entrydsm.rollsroyce.domain.admin.presentation.dto.request.CreateReplyRequest;
 import kr.hs.entrydsm.rollsroyce.domain.admin.presentation.dto.request.GetApplicantsRequest;
+import kr.hs.entrydsm.rollsroyce.domain.admin.presentation.dto.request.UpdateReplyRequest;
 import kr.hs.entrydsm.rollsroyce.domain.admin.presentation.dto.response.ApplicantsResponse;
 import kr.hs.entrydsm.rollsroyce.domain.admin.presentation.dto.response.StaticsCountResponse;
 import kr.hs.entrydsm.rollsroyce.domain.admin.presentation.dto.response.StaticsScoreResponse;
+import kr.hs.entrydsm.rollsroyce.domain.admin.service.CreateReplyService;
 import kr.hs.entrydsm.rollsroyce.domain.admin.service.DeleteAllTablesService;
 import kr.hs.entrydsm.rollsroyce.domain.admin.service.DeleteQnaService;
 import kr.hs.entrydsm.rollsroyce.domain.admin.service.GetApplicantsService;
 import kr.hs.entrydsm.rollsroyce.domain.admin.service.QueryStaticsCountService;
 import kr.hs.entrydsm.rollsroyce.domain.admin.service.QueryStaticsScore;
+import kr.hs.entrydsm.rollsroyce.domain.admin.service.UpdateReplyService;
 
 @RequiredArgsConstructor
 @RequestMapping("/admin")
@@ -34,6 +43,8 @@ public class AdminController {
     private final QueryStaticsCountService queryStaticsCountService;
     private final QueryStaticsScore queryStaticsScore;
     private final DeleteQnaService deleteQnaService;
+    private final CreateReplyService createReplyService;
+    private final UpdateReplyService updateReplyService;
 
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping("/data")
@@ -57,8 +68,20 @@ public class AdminController {
     }
 
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    @DeleteMapping("/qna/{qna-id}")
+    @DeleteMapping("/{qna-id}")
     public void deleteQna(@PathVariable("qna-id") Long qnaId) {
         deleteQnaService.execute(qnaId);
+    }
+
+    @ResponseStatus(HttpStatus.CREATED)
+    @PostMapping("/{qna-id}")
+    public void createReply(@RequestBody @Valid CreateReplyRequest request, @PathVariable("qna-id") Long qnaId) {
+        createReplyService.execute(request, qnaId);
+    }
+
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    @PatchMapping("/{reply-id}")
+    public void updateReply(@PathVariable("reply-id") Long replyId, @RequestBody @Valid UpdateReplyRequest request) {
+        updateReplyService.execute(replyId, request);
     }
 }
